Relocate the admin pages + Remove the preact build
