+++ src/main/javascript/src/pages/employees/update-employee.js
@@ -1,12 +1,12 @@
 import {Component} from 'preact';
-import Layout from "../../../components/layout";
-import RoleLock from "../../../components/role-lock";
+import Layout from "../../components/layout";
+import RoleLock from "../../components/role-lock";
 import {connect} from "preact-redux";
-import reducers from "../../../reducers";
-import * as actions from "../../../actions";
-import Error from "../../../components/error";
+import reducers from "../../reducers";
+import * as actions from "../../actions";
+import Error from "../../components/error";
 import {Link} from "preact-router/match";
-import CustomError from "../../../components/error-custom";
+import CustomError from "../../components/error-custom";
 
 @connect(reducers, actions)
 export default class UpdateEmployeePage extends Component {
